app/requirements.txt
AHRS==0.3.1
asgiref==3.5.2
APScheduler==3.10.1
bcrypt==4.0.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.3
Django==4.1.2
django-admin-logs==1.0.2
django-admin-rangefilter==0.12.4
django-cors-headers==3.13.0
django-rest-framework==0.1.0
django-sql-utils==0.6.1
djangorestframework==3.14.0
drf-yasg==1.21.4
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
paramiko==2.12.0
passlib==1.7.4
password-strength==0.0.3.post2
pika==1.3.1
psycopg2-binary==2.9.9
pycparser==2.21
PyNaCl==1.5.0
pyparsing==3.0.9
pyquaternion==0.9.9
pysftp==0.2.9
python-decouple==3.8
pytz==2022.6
requests==2.28.1
simplejwt==2.0.1
six==1.16.0
sqlparse==0.4.3
uritemplate==4.1.1
user-agents==2.2.0
urllib3==1.26.12
pymysql
pytz==2022.6
pillow
python-dateutil==2.9.0.post0
pytest-cov
pytest-django
coverage
gevent
django-db-geventpool==4.0.2
celery==5.5.2
redis==5.2.1
django-redis==5.4.0
pandas
sentence-transformers>=2.7.0
huggingface_hub>=0.14.1,<1.0.0
torch>=2.1